requirements.txt
altair==5.5.0
amqp==5.3.1
anyio==4.11.0
arabic-reshaper==3.0.0
asgiref==3.10.0
asn1crypto==1.5.1
attrs==25.3.0
autobahn==25.10.2
Automat==25.4.16
billiard==4.2.2
bleach==6.3.0
blinker==1.9.0
Brotli==1.1.0
cachetools==6.2.0
celery==5.5.3
certifi==2025.1.31
cffi==1.17.1
channels==4.3.2
channels_redis==4.3.0
chardet==5.2.0
charset-normalizer==3.4.1
choreographer==1.0.9
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.2
crispy-bootstrap5==2025.4
cryptography==46.0.0
cssselect2==0.8.0
cycler==0.12.1
daphne==4.2.1
defusedxml==0.7.1
dj-rest-auth==7.0.1
Django==5.2
django-appconf==1.1.0
django-bootstrap5>=24.3
django-bootstrap5==25.2
django-crispy-forms==2.4
django-encrypted-model-fields==0.6.5
django-extensions==4.1
django-htmx==1.26.0
django-localflavor==5.0
django-phonenumber-field==8.1.0
django-rest-framework==0.1.0
django-select2==8.4.3
django-sendfile2==0.7.2
django-signature==0.3.1
django-template-partials==25.1
django-widget-tweaks==1.5.0
django_components==0.141.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
djc_core_html_parser==1.0.2
dnspython==2.8.0
docopt==0.6.2
docx-mailmerge==0.5.0
docx-mailmerge2==0.9.0
et_xmlfile==2.0.0
eventlet==0.40.3
fonttools==4.56.0
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
honcho==2.0.0
html5lib==1.1
httptools==0.7.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
javascript==1!1.2.5
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kaleido==1.0.0
kiwisolver==1.4.8
kombu==5.5.4
logistro==1.1.0
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib==3.10.3
msgpack==1.1.2
mysqlclient==2.2.7
narwhals==1.47.1
npm==0.1.1
num2words==0.5.14
numpy==2.3.1
openpyxl==3.1.5
optional-django==0.1.0
orjson==3.11.0
oscrypto==1.3.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.0
phonenumbers==9.0.13
pillow==11.1.0
playwright==1.51.0
plotly==6.2.0
plyer==2.1.0
prompt_toolkit==3.0.52
protobuf==6.32.0
psutil==7.1.3
psycopg2==2.9.10
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydeck==0.9.1
pydyf==0.11.0
pyee==12.1.1
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
PyJWT==2.8.0
PyMySQL==1.1.2
pyOpenSSL==25.3.0
pyparsing==3.2.3
pypdf==5.4.0
pyphen==0.17.2
PySocks==1.7.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
python-dotenv==1.1.1
python-magic==0.4.27
python-magic-bin==0.4.14
python-stdnum==2.1
pytz==2025.2
PyYAML==6.0.2
qrcode==8.1
redis==6.4.0
referencing==0.36.2
reportlab==4.4.0
requests==2.32.3
rpds-py==0.27.1
selenium==4.30.0
service-identity==24.2.0
setuptools==80.9.0
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
streamlit==1.49.1
streamlit-drawable-canvas==0.9.3
streamlit-pdf-viewer==0.0.26
svglib==1.5.1
tenacity==9.1.2
tinycss2==1.4.0
tinyhtml5==2.0.0
toml==0.10.2
tornado==6.5.2
trio==0.29.0
trio-websocket==0.12.2
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.13.0
tzdata==2025.2
tzlocal==5.3.1
uritools==4.0.3
urllib3==2.3.0
uvicorn==0.38.0
vine==5.1.0
waitress==3.0.2
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.14
weasyprint==65.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
whitenoise==6.11.0
wsproto==1.2.0
xhtml2pdf==0.2.17
zope.interface==8.0.1
zopfli==0.2.3.post1
django-bootstrap5
